@@ -2,39 +2,56 @@
 
 buildscript {
     ext {
-        kotlinVersion = '1.2.30'
-        supportLibraryVersion = '27.1.0'
-        supportTestRunnerVersion = '1.0.1'
-        junitVersion = '4.1.2'
+        kotlinVersion = '1.4.30'
+        androidx = [
+                annotationsVersion : "1.1.0",
+                fragmentVersion    : "1.3.0",
+                recyclerViewVersion: "1.1.0",
+                test               : [
+                        junitVersion : "1.1.2",
+                        runnerVersion: "1.3.0"
+                ]
+        ]
     }
     repositories {
         google()
+        mavenCentral()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:4.2.0-beta04'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
     }
+}
+
+plugins {
+    id "com.diffplug.spotless" version "5.10.1"
 }
 
 allprojects {
     repositories {
         google()
+        mavenCentral()
         jcenter()
     }
 }
 
-task clean(type: Delete) {
-    delete rootProject.buildDir
+subprojects {
+    apply plugin: 'com.diffplug.spotless'
+    spotless {
+        kotlin {
+            target("**/*.kt")
+            targetExclude("$buildDir/**/*.kt")
+            ktlint("0.40.0")
+        }
+    }
 }
 
 ext {
-    compileSdkVersion = 27
-    buildToolsVersion = '27.0.3'
+    compileSdkVersion = 30
 
     minSdkVersion = 16
-    targetSdkVersion = 27
-    versionCode = 16
-    versionName = "0.10.0"
+    targetSdkVersion = 30
+    versionCode = 19
+    versionName = "1.0.1"
 }
