@@ -3,26 +3,13 @@ package com.github.droibit.chopstick.preference
 import android.preference.Preference
 import android.preference.PreferenceActivity
 import android.preference.PreferenceFragment
-import android.support.annotation.StringRes
-import android.support.v7.preference.PreferenceFragmentCompat
-import android.support.v14.preference.PreferenceFragment as PreferenceFragmentV14
-import android.support.v7.preference.Preference as PreferenceV7
+import androidx.annotation.StringRes
+import androidx.preference.PreferenceFragmentCompat
+import kotlin.LazyThreadSafetyMode.NONE
+import androidx.preference.Preference as PreferenceCompat
 
-fun <P : PreferenceV7> PreferenceFragmentCompat.bindPreference(@StringRes resId: Int): Lazy<P> =
-  requireV7 { findPreference(getString(resId)) }
-
-fun <P : PreferenceV7> PreferenceFragmentCompat.bindPreference(key: String): Lazy<P> =
-  requireV7 { findPreference(key) }
-
-fun <P : PreferenceV7> PreferenceFragmentV14.bindPreference(@StringRes resId: Int): Lazy<P> =
-  requireV7 { findPreference(getString(resId)) }
-
-fun <P : PreferenceV7> PreferenceFragmentV14.bindPreference(key: String): Lazy<P> =
-  requireV7 { findPreference(key) }
-
-@Suppress("UNCHECKED_CAST")
-private inline fun <P : PreferenceV7> requireV7(crossinline finder: () -> PreferenceV7?) =
-  lazy { requireNotNull(finder()) as P }
+fun <P : PreferenceCompat> PreferenceFragmentCompat.bindPreference(@StringRes resId: Int): Lazy<P> =
+    requiredCompat { findPreference(getString(resId)) }
 
 @Suppress("DEPRECATION")
 @Deprecated(
@@ -31,7 +18,10 @@ private inline fun <P : PreferenceV7> requireV7(crossinline finder: () -> Prefer
     level = DeprecationLevel.WARNING
 )
 fun <P : Preference> PreferenceActivity.bindPreference(@StringRes resId: Int): Lazy<P> =
-  require { findPreference(getString(resId)) }
+    required { findPreference(getString(resId)) }
+
+fun <P : PreferenceCompat> PreferenceFragmentCompat.bindPreference(key: String): Lazy<P> =
+    requiredCompat { findPreference(key) }
 
 @Suppress("DEPRECATION")
 @Deprecated(
@@ -40,14 +30,18 @@ fun <P : Preference> PreferenceActivity.bindPreference(@StringRes resId: Int): L
     level = DeprecationLevel.WARNING
 )
 fun <P : Preference> PreferenceActivity.bindPreference(key: String): Lazy<P> =
-  require { findPreference(key) }
+    required { findPreference(key) }
 
 fun <P : Preference> PreferenceFragment.bindPreference(@StringRes resId: Int): Lazy<P> =
-  require { findPreference(getString(resId)) }
+    required { findPreference(getString(resId)) }
 
 fun <P : Preference> PreferenceFragment.bindPreference(key: String): Lazy<P> =
-  require { findPreference(key) }
+    required { findPreference(key) }
 
 @Suppress("UNCHECKED_CAST")
-private inline fun <P : Preference> require(crossinline finder: () -> Preference?) =
-  lazy { requireNotNull(finder()) as P }
+private inline fun <P : Preference> required(crossinline finder: () -> Preference?): Lazy<P> =
+    lazy(NONE) { requireNotNull(finder()) as P }
+
+@Suppress("UNCHECKED_CAST")
+private inline fun <P : PreferenceCompat> requiredCompat(crossinline finder: () -> PreferenceCompat?) =
+    lazy(NONE) { requireNotNull(finder()) as P }
