@@ -14,30 +14,30 @@ import com.github.droibit.chopstick.sample.R
 @Suppress("DEPRECATION")
 class SettingsActivityCompat : AppCompatPreferenceActivity() {
 
-  companion object {
+    companion object {
 
-    @JvmStatic
-    fun makeIntent(context: Context) = Intent(context, SettingsActivityCompat::class.java)
-  }
-
-  private val checkboxPref: CheckBoxPreference by bindPreference(R.string.key_checkbox_preference)
-  private val switchPref: SwitchPreference by bindPreference("switch_preference")
-  private val editPref: EditTextPreference by bindPreference(R.string.key_edittext_preference)
-  private val listPref: Preference by bindPreference("list_preference")
-
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    addPreferencesFromResource(R.xml.settings)
-
-    checkboxPref.apply {
-      isChecked = false
-      title = "Edited: ${checkboxPref.title}"
+        @JvmStatic
+        fun makeIntent(context: Context) = Intent(context, SettingsActivityCompat::class.java)
     }
-    switchPref.apply {
-      isChecked = false
-      title = "Edited: ${switchPref.title}"
+
+    private val checkboxPref: CheckBoxPreference by bindPreference(R.string.key_checkbox_preference)
+    private val switchPref: SwitchPreference by bindPreference("switch_preference")
+    private val editPref: EditTextPreference by bindPreference(R.string.key_edittext_preference)
+    private val listPref: Preference by bindPreference("list_preference")
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        addPreferencesFromResource(R.xml.settings)
+
+        checkboxPref.apply {
+            isChecked = false
+            title = "Edited: ${checkboxPref.title}"
+        }
+        switchPref.apply {
+            isChecked = false
+            title = "Edited: ${switchPref.title}"
+        }
+        editPref.title = "Edited: ${editPref.title}"
+        listPref.title = "Edited: ${listPref.title}"
     }
-    editPref.title = "Edited: ${editPref.title}"
-    listPref.title = "Edited: ${listPref.title}"
-  }
 }
