@@ -1,30 +1,31 @@
 package com.github.droibit.chopstick.sample.prefs.fragment
 
 import android.os.Bundle
-import android.preference.CheckBoxPreference
-import android.preference.EditTextPreference
-import android.preference.Preference
-import android.preference.PreferenceFragment
-import android.preference.SwitchPreference
 import android.view.View
+import androidx.preference.CheckBoxPreference
+import androidx.preference.EditTextPreference
+import androidx.preference.Preference
+import androidx.preference.PreferenceFragmentCompat
+import androidx.preference.SwitchPreferenceCompat
 import com.github.droibit.chopstick.preference.bindPreference
 import com.github.droibit.chopstick.sample.R
 
-class SettingsFragment : PreferenceFragment() {
+class SettingsFragmentCompat : PreferenceFragmentCompat() {
 
     private val checkboxPref: CheckBoxPreference by bindPreference(R.string.key_checkbox_preference)
-    private val switchPref: SwitchPreference by bindPreference("switch_preference")
+    private val switchPref: SwitchPreferenceCompat by bindPreference("switch_preference")
     private val editPref: EditTextPreference by bindPreference(R.string.key_edittext_preference)
     private val listPref: Preference by bindPreference("list_preference")
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        addPreferencesFromResource(R.xml.settings)
+    override fun onCreatePreferences(
+        savedInstanceState: Bundle?,
+        rootKey: String?
+    ) {
+        addPreferencesFromResource(R.xml.settings_androidx)
     }
 
     override fun onViewCreated(
-        view: View?,
+        view: View,
         savedInstanceState: Bundle?
     ) {
         super.onViewCreated(view, savedInstanceState)
